Use zerocopy for scalar {un}marshalling.
Keeps wrapping trait b/c we will want it for structs with variable
sized encoding, but this allows us to use zerocopy for primitive types.

Also tweaks the try_marshal interface to use a buffer rather than return
a vector. This allows a similar mirroring of the zerocopy AsBytes
api.

Signed-off-by: Emily Musick <emusick@google.com>