Full TpmtPublic and TpmlPcrSelection {un}marshaling examples (#17)
This exercises full (and I believe correct!) {un}marshaling for TpmtPublic and TpmlPcrSelection structs as examples that feature most interesting TPM marshaling cases.

It uses a procedural macro to generate most of the marshaling code for structs that cannot derive zerocopy::{AsBytes,FromBytes}, with attributes for handling union selectors and array-count fields. Union types implement the Marshalable trait themselves to provide selector->variant mappings. The macro and its use are currently limited to the subset of what I needed to get my example types to work, but can be expanded/applied to everything once we're happy with them.
There are of course plenty of improvements, and in particular I haven't focused much on optimizing anything, but my goal is to provide an idea of how to generally/scalably implement {un}marshaling for TPM structs.

Signed-off-by: Emily Musick <emusick@google.com>