Replace unions with enums in example types (#22)
This switches from using unions to using enums for types with marshaling support. This trades full c-layout compatibility and adds a bit more complexity in the marshal macro for no longer needing unsafe to access union variants. A mismatch between selector and variant is no longer possible, and we can generate enum marshaling routines with the macro.

TPMT_PUBLIC now requires a manual implementation of marshaling routines instead. I think this is the right tradeoff, because it's doing the somewhat "weird" thing. We could, of course, add some way to express this in the macro as well, but that seems premature at this point.

Signed-off-by: Emily Musick <emusick@google.com>