Extend marshal approach to all types.
Use BE types and enums with fields throughout. This was done for many
types to get a couple of examples e2e {un}marshaling, but this makes it
consistent. Unions only used for size info are the exception because I'm
not yet sure how we want to build those.

Also derive PartialEq and Debug everywhere.

Signed-off-by: Emily Musick <emusick@google.com>