Extend marshal approach to all types. (#24)
Use BE types and enums with fields throughout. This was done for many
types to get a couple of examples e2e {un}marshaling, but this makes it
consistent. Unions only used for size info compute and the relevant size in 
const fn union_size(). 

Also derive PartialEq and Debug everywhere.

Signed-off-by: Emily Musick <emusick@google.com>