Replace type aliases with new types
Replace all types aliasing to an integer with a struct wrapping that
integer. This makes it far harder to make a mistake when using these
types since the type checker will now catch it if we accidentally use
Tpm2AlgId in place of a TpmiAlgSymMode for example.

Additionally, this has 0 additional memory footprint.