Perform BE conversion only at marshaling time. (#38)
Since not everything is stored in wire format, we get little benefit
from using zerocopy types. This switches everything to use normal Rust
primitives and performs BE conversion to {un}marshal.

We aim to keep this change "reversible" so we can swap out the marshaling primitives to test the impact of BE storage on efficiency and code size.