Replace repr(C) with transparent for tuple structs (#49)
This patch modifies all structs in the form of `struct Foo(Bar)` to be
represented transparently instead of being represented as C structure,
hence fixing a potential undefined behaviour.

This allows C-bindings to expect the inner datatype without being
wrapped inside another struct. This might lead to some subtle ABI
differences. For instance, some architectures treat C structs
differently from integers for instance.

The other alternative is to replace `struct Foo(Bar)` with `struct
Foo{inner: Bar}`, but there is no advantage in forcing C users to wrap
every single integer value in a new structure.

For reference: According to Rust Function call ABI compatibility[1]:
```
if you have a C function that is defined to take a uint32_t:
        void some_function(uint32_t value) { .. }
It is incorrect to pass in a struct as that value, even if that struct is #[repr(C)] and has only one field:
        #[repr(C)]
        struct Foo { x: u32 }
        extern "C" some_function(Foo);
        some_function(Foo { x: 22 }); // Bad!
Instead, you should declare the struct with #[repr(transparent)], which specifies that Foo should use the ABI rules for its field type, u32.
...
...
(Note further that the Rust ABI is undefined and theoretically may vary from compiler revision to compiler revision.)
```

[1]
https://rust-lang.github.io/unsafe-code-guidelines/layout/structs-and-tuples.html

Signed-off-by: Ahmed Abdelraoof <ahmed.abdelraoof@huawei.com>