Add examples for richer Tpmi* and Tpma* types.
Initially I was hoping to use a fieldless enum for Tpmi types, however
we do want to be able to use this in FFI to C, which would be unsafe.
So, I use another open_enum, which will still allow cleaner usage. It
will also work for the TPMI*+ types without a separate definition.

I also added an example TPMA* type that will allow construction from a
set of flags. This too is FFI compatible.