Fix union_size calculation for Tpmu* and Tpmt*
In particular this patch uses the `unionify` macro to construct a union
version of each enum, thus delegating the calculation of the union size
to the compiler. This patch unionifies the following types:
- TpmtHa
- TpmuName
- TpmuSensitiveCreate
- TpmuEncryptedSecret

It is not entirely impossible that there are more hidden instances of
the same bug. I only searched for instances of `align_of` in the whole
code base and used that as my only cue when looking for bug instanes.

Signed-off-by: Ahmed Abdelraoof <ahmed.abdelraoof@huawei.com>