FIX #55: Delegate size calculation for TPMU* to rust compiler (#62)
This patch implements `UnionSize` a trait + derive macro. The trait
contains a constans representing the size of union equivalent to the
enum in question.The derive macro generates a union from the enum in
question, and use it to implement the trait in question.

Then we use this  macro to construct a union version of each enum,
thus delegating the calculation of the union size to the compiler. This patch
unionifies the following types:
- TpmtHa
- TpmuName
- TpmuSensitiveCreate
- TpmuEncryptedSecret

It is not entirely impossible that there are more hidden instances of
the same bug. I only searched for instances of `align_of` in the whole
code base and used that as my only cue when looking for bug instanes
---------

Signed-off-by: Ahmed Abdelraoof <ahmed.abdelraoof@huawei.com>