[marshal/derive] Add support for marshallig refinement types
This patch extends the capability of the Marshal macro by adding
attribute #[marshal(via=T)] which allows a type to be marshalled by
converting it to another type before {un_}marshalling.

This allows for more expressive types all while having abstracting all
the runtime checks behind `TryInto`

Example: Imagine we have the following:

    #[marshal(via = TpmHc)]
    pub enum TpmiDhObject {
        Transient { index: u24 } = TpmHc::HRTransient.get(),
        Persistent { index: u24 } = TpmHc::HRPersistent.get(),
    }

This will expand to:
    impl Marshalable for TpmiDhObject  {
        fn try_unmarshal(buffer: &mut UnmarshalBuf) -> tpm2_rs_marshal::exports::errors::TpmRcResult<Self> {
            let unrefined = TpmHc::try_unmarshal(buffer)?;
            unrefined.try_into()
        }
        fn try_marshal(&self, buffer: &mut [u8]) -> tpm2_rs_marshal::exports::errors::TpmRcResult<usize> {
            let unrefined : TpmHc = self.into();
            unrefined.try_marshal(buffer)
        }
    }

It is the macro's user responsibility to implement the necessary
From/Into and TryFrom/TryInto to make this conversion happen. But
on the flip side, users get more natural marshaling api for interface
types.

Signed-off-by: Ahmed Abdelraoof <ahmed.abdelraoof@huawei.com>