Separate unit tests into a different file when there are many of them
This change will make it easier to manage our unit tests.

There is nothing special about the `test` module in rust libraries, it's just a separate module that is conditionally compiled based on whether or not the crate is running with the `test` command. As such, it's possible (and allowable) to separate the unit test code into a separate file.

It would probably be wise to move these tests closer to the code that they test, but let's do that in the future.

There is one exception made to this layout:

*   Tests written under the `tests/` directory (outside of the `src` directory), are special. Each `*.rs` file is compiled as a different crate, and then the crate has tests run on it. This allows us to test the usage of our crate as-if we are an outside client. However, these tests don't allow access to private data or functions, so it's not always the most useful thing to use for all testing.
*   The only thing we use integration tests for currently are simulator tests. We don't want the simulator tests to run for any random invocation of `test` - because the client may not have the simulator running. So we hide those tests in a submodule and only include it if the feature for the tests is enabled.
*   Because of this, we cannot define the tests in a leaf `simulator_tests.rs` file, otherwise it will be compiled and run since it's in the root of `tests/` - so we hide it in a submodule in the format `simulator_tests/mod.rs`.
*   This PR is not adding this newly, it's just cleaning up what is there to unnest the tests a bit and make it more obvious what is happening.