Rename the tpm2-rs-marshal and tpm2-rs-marshal-derive
A prior PR updated the `Marshal` proc-macro to be `Marshalable` in order to make the naming more consistent. However this creates a strange naming standard for the crate - which was originally named `tpm2-rs-marshal` and `tpm2-rs-marshal-derive`. This commit renames the crates to match with the trait and proc-macro that the crates are providing.