Rename tpm2-rs-marshal and tpm2-rs-marshal-derive (#81)
A prior PR (https://github.com/tpm-rs/tpm-rs/pull/80) updated the
`Marshal` proc-macro to be `Marshalable` in order to make the naming
more consistent. However this creates a strange naming standard for the
crate - which was originally named `tpm2-rs-marshal` and
`tpm2-rs-marshal-derive`. This commit renames the crates to match with
the trait and proc-macro that the crates are providing.