Add --tests flag to cargo clippy CI
In doing so, this patch fixes couple of clippy warnings. They all fall
in one of those categories:

- unneeded `return` statement
- field assignment outside of initializer for an instance created with
  Default::default() (essentially doing a let binding only to modify on
  the next line)
- unnecessary double references
- unnecessary cloning
- unnecessary `.into()`

There is however one warning that we chose to ignore, which is defining
a unit field in a structure. We ignore that warning because we are
testing the behavior of marshalable on zero sized types.

It was hard to convince clippy to ignore this warning and the only way
that worked was placing the `allow` module-wide.

fixes: https://github.com/tpm-rs/tpm-rs/issues/77

Signed-off-by: Ahmed Abdelraoof <ahmed.abdelraoof@huawei.com>