Add a new trait: MarshalableVariant
This change starts the process of implementing a separate trait
for marshaling variant data independently of the discriminant.
This PR purposefully aims to do the minimal amount of changes
so that we can unblock other cleanup in other PRs.

After this PR we will do more cleanup in the repo, and eventually
we will separate this out into its own derive proc-macro. But
to unblock the cleanup of tpm2-rs-base, we need to start with this.

See the discussion about this: https://github.com/tpm-rs/tpm-rs/issues/84

Changes:

*   Added a new trait named `MarshalableVariant`.
*   Added a single test in `tpm2-rs-marshalable`, we can do better here
    but I'd like to refactor the error codes first, because right now it's
    hard to work with. So for now, since I'm not making any functional
    changes, I'd like to do the bare minimum in testing (we'll add more later).