Move unsafe discriminant calculation outside tpm-rs
This is following discussions in https://github.com/tpm-rs/tpm-rs/issues/79
This patch dependents on the crate safe-discriminant instead of
calculating discriminants on the fly. Objectively, safe-discriminant is not
safer than the solution we have. But, it isolates the unsafe code in its
separate crate making it easier to review the code and evaluate its safety.

fixes #79

Signed-off-by: Ahmed Abdelraoof <ahmed.abdelraoof@huawei.com>