server: Rename tpm2-rs-service to tpm2-rs-server
Even in the meeting discussion I often found myself and others calling
that part of the project the "server" given that the other part is
called the client. Naming it service is a a little bit confusing hence I
propose officially calling it server instead.

This patch does not modify any business logic anywhere.

Signed-off-by: Ahmed Abdelraoof <ahmed.abdelraoof@huawei.com>