server: Eliminate Crypto lifetime from TpmContext
The modification in this patch simplifis
`TpmContext<'a, Deps: TpmContextDeps>` into `TpmContext<Deps: TpmContextDeps>`
The lifetime was meant for the crypto engine. This simplifies the implementation
and avoid 1 extra level of indirection. If the real physical crypto resource
(HSM engine) is not owned by the TPMContext, it is the implementer responsibility
to abstract over that and make sure the provided crypto resource lives long enough.

Signed-off-by: Ahmed Abdelraoof <ahmed.abdelraoof@huawei.com>