server: Delegate Crypto ownership to CommandHandler
In theory, TpmContext should not do any cryptographic operation on its
own. Currently it always constructs a short lived CommandHandler and let
it handle the crypographic operation on its behalf. This patch instead
creates a long lived CommandHandler and makes it own the cryptographic
context and makes it part of TpmContext. The main advantage is it avoids
construcing objects unnecesarily.

Signed-off-by: Ahmed Abdelraoof <ahmed.abdelraoof@huawei.com>