server: Hide unnecessary modules (#131)
In principle users need to be only aware of two parts of the server:
- The TpmContext structure which is the top level tpm implementation. It
consumes byte-array-like structures representing tpm commands and
produces byte-array-like structures representing tpm responses.
- The plaform module which represent the HAL layer.

Everything else can be hidden. In doing so, we discovered one method the
`Response::write()` which no-one uses so far. So we make overwrite lints
for that one function to allow it being deadcode.

Signed-off-by: Ahmed Abdelraoof <ahmed.abdelraoof@huawei.com>